ESSIFI-104 WIP initial impl for chaining input evaluations
