--- lib/evaluation/bundlers/index.ts
@@ -1,2 +1,0 @@
-export * from './evaluationBundler';
-export * from './candidateInputFitsInputDescriptorEB';
