--- lib/evaluation/core/evaluated.ts
@@ -1,3 +1,0 @@
-import { Checked, NonEmptyArray } from '../../ConstraintUtils';
-
-export type Evaluated = NonEmptyArray<Checked> | Checked;
