--- lib/evaluation/core/index.ts
@@ -1,2 +1,0 @@
-export * from './evaluated';
-export * from './evaluator';
