--- lib/evaluation/index.ts
@@ -1,3 +1,0 @@
-export * from './core';
-export * from './bundlers';
-export * from './evaluationEngine';
